+++ R/ZenodoRequest.R
@@ -14,6 +14,12 @@ ZenodoRequest <- R6Class("ZenodoRequest",
   inherit = zen4RLogger,                    
   #private methods
   private = list(
+    agent = paste(
+      "Mozilla/5.0 (X11; Linux x86_64)",
+      "AppleWebKit/537.36 (KHTML, like Gecko)",
+      "Chrome/103.0.0.0 Safari/537.36"
+    ),
+    #agent = paste0("zen4R", as(packageVersion("zen4R"), "character")),
     url = NA,
     type = NA,
     request = NA,
@@ -57,7 +63,10 @@ ZenodoRequest <- R6Class("ZenodoRequest",
     GET = function(url, request){
       req <- paste(url, request, sep="/")
       self$INFO(sprintf("Fetching %s", req))
-      headers <- c("Authorization" = paste("Bearer",private$token))
+      headers <- c(
+        "Authorization" = paste("Bearer",private$token),
+        "User-Agent" = private$agent
+      )
       
       r <- NULL
       if(self$verbose.debug){
@@ -82,8 +91,11 @@ ZenodoRequest <- R6Class("ZenodoRequest",
       }
       
       #headers
-      headers <- c("Content-Type" = contentType,
-                   "Authorization" = paste("Bearer",private$token))
+      headers <- c(
+        "Content-Type" = contentType,
+        "Authorization" = paste("Bearer",private$token),
+        "User-Agent" = private$agent
+      )
       
       #send request
       if(self$verbose.debug){
@@ -114,8 +126,11 @@ ZenodoRequest <- R6Class("ZenodoRequest",
       if(regexpr("api/files", req)<0) data <- private$prepareData(data)
       
       #headers
-      headers <- c("Content-Type" = if(regexpr("api/files", req)>0) "application/octet-stream" else "application/json",
-                   "Authorization" = paste("Bearer",private$token))
+      headers <- c(
+        "Content-Type" = if(regexpr("api/files", req)>0) "application/octet-stream" else "application/json",
+        "Authorization" = paste("Bearer",private$token),
+        "User-Agent" = private$agent
+      )
       
       #send request
       if(self$verbose.debug){
@@ -141,7 +156,10 @@ ZenodoRequest <- R6Class("ZenodoRequest",
     DELETE = function(url, request, data){
       req <- paste(url, request, data, sep="/")
       #headers
-      headers <- c("Authorization" = paste("Bearer",private$token))
+      headers <- c(
+        "Authorization" = paste("Bearer",private$token),
+        "User-Agent" = private$agent
+      )
       if(self$verbose.debug){
         r <- with_verbose(httr::DELETE(
           url = req,